Misc repairs

+++ classes/SpecProcessorOcr.php
@@ -312,31 +312,38 @@ class SpecProcessorOcr{
 		if(file_exists($zipPath)) {
 			unlink($zipPath);
 		}
-		if(move_uploaded_file($_FILES['ocrfile']['tmp_name'], $zipPath)){
-			$zip = new ZipArchive;
-			$res = $zip->open($zipPath);
-			if($res === true) {
-				$extractPath = $this->tempPath.'ocrtext_'.time().'/';
-				if (!mkdir($extractPath) && !is_dir($extractPath)) {
-					throw new RuntimeException(sprintf('Directory "%s" was not created', $extractPath));
-				}
-				if($zip->extractTo($extractPath)){
-					$retPath = $extractPath;
-				}
-				$zip->close();
-				unlink($zipPath);
-			}
-			else{
-				$this->errorStr = 'ERROR unpacking OCR file: '.$res;
-				$this->logMsg($this->errorStr);
-				return false;
-			}
-		}
+		if(is_writable($this->tempPath)){
+            if(move_uploaded_file($_FILES['ocrfile']['tmp_name'], $zipPath)){
+                $zip = new ZipArchive;
+                $res = $zip->open($zipPath);
+                if($res === true) {
+                    $extractPath = $this->tempPath.'ocrtext_'.time().'/';
+                    if (!mkdir($extractPath) && !is_dir($extractPath)) {
+                        throw new RuntimeException(sprintf('Directory "%s" was not created', $extractPath));
+                    }
+                    if($zip->extractTo($extractPath)){
+                        $retPath = $extractPath;
+                    }
+                    $zip->close();
+                    unlink($zipPath);
+                }
+                else{
+                    $this->errorStr = 'ERROR unpacking OCR file: '.$res;
+                    $this->logMsg($this->errorStr);
+                    return false;
+                }
+            }
+            else{
+                $this->errorStr = 'ERROR loading OCR file: input file lacks zip extension';
+                $this->logMsg($this->errorStr);
+                return false;
+            }
+        }
 		else{
-			$this->errorStr = 'ERROR loading OCR file: input file lacks zip extension';
-			$this->logMsg($this->errorStr);
-			return false;
-		}
+            $this->errorStr = 'ERROR loading OCR file: Destination is not writable to server';
+            $this->logMsg($this->errorStr);
+            return false;
+        }
 		return $retPath;
 	}
